+++ src/services/deviationsFeeds/DeviationTrigger.ts
@@ -1,7 +1,7 @@
 import {Logger} from 'winston';
 import {inject, injectable} from 'inversify';
 
-import {DeviationDataToSign, DeviationLeavesAndFeeds} from '../../types/DeviationFeeds.js';
+import {DeviationLeavesAndFeeds, DeviationTriggerResponse} from '../../types/DeviationFeeds.js';
 import {PriceDataProvider} from './PriceDataProvider.js';
 import {DeviationTriggerFilters} from './DeviationTriggerFilters.js';
 import {DataFilter} from '../tools/DataFilter.js';
@@ -19,12 +19,12 @@ export class DeviationTrigger {
     dataTimestamp: number,
     data: DeviationLeavesAndFeeds,
     pendingChains?: Set<string>,
-  ): Promise<DeviationDataToSign | undefined> {
+  ): Promise<DeviationTriggerResponse> {
     const {feeds, leaves} = data;
 
     if (leaves.length === 0) {
       this.logger.debug(`[DeviationTrigger] no leaves to process for ${dataTimestamp}`);
-      return;
+      return {reason: '0 leaves/prices to process'};
     }
 
     // discard feeds for which we do not have leaves
@@ -33,8 +33,11 @@ export class DeviationTrigger {
       leaves.map((l) => l.label),
     );
 
+    const reasons: string[] = [];
+
     if (removed.length != 0) {
       this.logger.warn(`[DeviationTrigger] removed feeds (no price): ${removed}`);
+      reasons.push(`removed feeds (no price): ${removed}`);
     }
 
     this.logger.info(`[DeviationTrigger] feeds: ${Object.keys(feeds)}`);
@@ -43,7 +46,7 @@ export class DeviationTrigger {
 
     const priceDataPerChain = await this.priceDataProvider.apply(keysPerChain);
 
-    const dataToUpdate = await this.deviationTriggerFilters.apply(
+    const triggerResponse = await this.deviationTriggerFilters.apply(
       dataTimestamp,
       leaves,
       feeds,
@@ -51,14 +54,23 @@ export class DeviationTrigger {
       pendingChains,
     );
 
-    if (!dataToUpdate || Object.keys(dataToUpdate.proposedPriceData).length == 0) {
+    if (triggerResponse.reason) {
+      reasons.push(triggerResponse.reason);
+    }
+
+    if (!triggerResponse.dataToUpdate || Object.keys(triggerResponse.dataToUpdate.proposedPriceData).length == 0) {
       this.logger.info(`[DeviationTrigger] no data to update for ${dataTimestamp}`);
-      return;
+      return {reason: `no data to update: ${reasons.join(';')}`};
     }
 
     // TODO in future we might want to remove the invalid data, atm we allow encoder to throw error
-    Object.values(dataToUpdate.proposedPriceData).forEach((data) => this.priceDataOverflowChecker.apply(data));
+    Object.values(triggerResponse.dataToUpdate.proposedPriceData).forEach((data) =>
+      this.priceDataOverflowChecker.apply(data),
+    );
 
-    return dataToUpdate;
+    return {
+      dataToUpdate: triggerResponse.dataToUpdate,
+      reason: reasons.join(';'),
+    };
   }
 }