+++ src/types/Settings.ts
@@ -175,6 +175,7 @@ type Settings = {
       apiKey: string;
     };
     uniswap: {
+      active: boolean;
       scannerContractId: string;
       helperContractId: string;
       startBlock: number;